+++ nextjs_moralis_nft_gating/package.json
@@ -9,7 +9,7 @@
         "lint": "next lint"
     },
     "dependencies": {
-        "axios": "^0.27.2",
+        "@moralisweb3/next": "^2.8.0",
         "moralis": "^2.8.0",
         "next": "^12.2.4",
         "next-auth": "^4.10.3",
@@ -17,5 +17,9 @@
         "react-dom": "^18.2.0",
         "wagmi": "^0.6.1",
         "ethers": "^5.6.9"
+    },
+    "devDependencies": {
+        "@types/react": "18.0.25",
+        "typescript": "4.9.3"
     }
 }
