+++ nextjs_moralis_nft_gating/pages/api/moralis/[...moralis].ts
@@ -1,0 +1,10 @@
+import { MoralisNextApi } from '@moralisweb3/next';
+
+export default MoralisNextApi({
+    apiKey: process.env.MORALIS_API_KEY,
+    authentication: {
+        domain: 'amazing.dapp',
+        uri: process.env.NEXTAUTH_URL,
+        timeout: 120,
+    },
+});
