+++ nextjs_moralis_nft_gating/pages/signin.jsx
@@ -1,14 +1,15 @@
-import { InjectedConnector } from 'wagmi/connectors/injected';
+import { MetaMaskConnector } from 'wagmi/connectors/metaMask';
 import { signIn } from 'next-auth/react';
 import { useAccount, useConnect, useSignMessage, useDisconnect } from 'wagmi';
 import { useRouter } from 'next/router';
-import axios from 'axios';
+import { useAuthRequestChallengeEvm } from '@moralisweb3/next';
 
 function SignIn() {
     const { connectAsync } = useConnect();
     const { disconnectAsync } = useDisconnect();
     const { isConnected } = useAccount();
     const { signMessageAsync } = useSignMessage();
+    const { requestChallengeAsync } = useAuthRequestChallengeEvm();
     const { push } = useRouter();
 
     const handleAuth = async () => {
@@ -16,17 +17,9 @@ function SignIn() {
             await disconnectAsync();
         }
 
-        const { account, chain } = await connectAsync({ connector: new InjectedConnector() });
+        const { account, chain } = await connectAsync({ connector: new MetaMaskConnector() });
 
-        const userData = { address: account, chain: chain.id, network: 'evm' };
-
-        const { data } = await axios.post('/api/auth/request-message', userData, {
-            headers: {
-                'content-type': 'application/json',
-            },
-        });
-
-        const message = data.message;
+        const { message } = await requestChallengeAsync({ address: account, chainId: chain.id });
 
         const signature = await signMessageAsync({ message });
 
@@ -42,7 +35,7 @@ function SignIn() {
     return (
         <div>
             <h3>Web3 Authentication</h3>
-            <button onClick={() => handleAuth()}>Authenticate via Metamask</button>
+            <button onClick={handleAuth}>Authenticate via Metamask</button>
         </div>
     );
 }
